pulled up emitTelemetryForCertBasedAuthResults
