+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/AzureActiveDirectoryWebViewClient.java
@@ -41,8 +41,8 @@ import androidx.annotation.RequiresApi;
 import com.microsoft.identity.common.adal.internal.AuthenticationConstants;
 import com.microsoft.identity.common.adal.internal.util.StringExtensions;
 import com.microsoft.identity.common.internal.broker.PackageHelper;
+import com.microsoft.identity.common.internal.ui.webview.certbasedauth.AbstractCertBasedAuthChallengeHandler;
 import com.microsoft.identity.common.internal.ui.webview.certbasedauth.CertBasedAuthFactory;
-import com.microsoft.identity.common.internal.ui.webview.certbasedauth.ICertBasedAuthChallengeHandler;
 import com.microsoft.identity.common.java.ui.webview.authorization.IAuthorizationCompletionCallback;
 import com.microsoft.identity.common.java.challengehandlers.PKeyAuthChallenge;
 import com.microsoft.identity.common.java.challengehandlers.PKeyAuthChallengeFactory;
@@ -80,7 +80,7 @@ public class AzureActiveDirectoryWebViewClient extends OAuth2WebViewClient {
     public static final String ERROR_DESCRIPTION = "error_description";
     private final String mRedirectUrl;
     private final CertBasedAuthFactory mCertBasedAuthFactory;
-    private ICertBasedAuthChallengeHandler mCertBasedAuthChallengeHandler;
+    private AbstractCertBasedAuthChallengeHandler mCertBasedAuthChallengeHandler;
 
     public AzureActiveDirectoryWebViewClient(@NonNull final Activity activity,
                                              @NonNull final IAuthorizationCompletionCallback completionCallback,
@@ -464,7 +464,7 @@ public class AzureActiveDirectoryWebViewClient extends OAuth2WebViewClient {
         }
         mCertBasedAuthFactory.createCertBasedAuthChallengeHandler(new CertBasedAuthFactory.CertBasedAuthChallengeHandlerCallback() {
             @Override
-            public void onReceived(@Nullable final ICertBasedAuthChallengeHandler challengeHandler) {
+            public void onReceived(@Nullable final AbstractCertBasedAuthChallengeHandler challengeHandler) {
                 mCertBasedAuthChallengeHandler = challengeHandler;
                 if (mCertBasedAuthChallengeHandler == null) {
                     //User cancelled out of CBA.
